requirements.txt
langchain-google-genai
langchain-core
python-dotenv
PyMuPDF